fix: start the update timeout, drop the CWD qt fallback, publish the crash breadcrumb atomically
Four small verified fixes:

- UpdateChecker armed a 10 s QTimer without start() and set no
  transfer timeout, so a connected-but-silent server hung the nested
  event loop forever (an invisible zombie in auto mode). Both bounds
  are live now: timer.start() plus setTransferTimeout(10 s).
- Executable-path resolution existed in five copies with four
  different truncation semantics; every call site now goes through
  pathutil::exePath()/exeDirectory() (truncation = failure), and the
  CWD-relative addLibraryPath("qt") fallback, reachable by all three
  GUI apps including elevated Editor runs, is deleted (fail closed to
  Qt's default search).
- Two %d-receives-size_t format mismatches (ChannelLayout,
  GraphicEQFilterFactory) fixed with %u + static_cast<unsigned>.
- The crash breadcrumb was a plain wchar_t[256] racing the faulting
  thread; a torn NUL could send %s past the buffer. Now a double
  buffer with an atomic index: writer fills the inactive slot and
  release-publishes, reader takes a bounded snapshot with one retry
  and a fixed fallback. Verified live: --selftest-crash produced the
  dump and report with the breadcrumb through the new path.

EditorLogicTests 3147, EngineOrchestrationTests 1317, HybridConvTests
1635 checks green locally.

+++ audio/ChannelLayout.cpp
@@ -118,7 +118,7 @@ int ChannelLayout::getChannelIndex(std::wstring word, const std::vector<std::wst
 
 		if (channelIndex < 0 || channelIndex >= static_cast<int>(channelNames.size()))
 		{
-			LogFStatic(L"Channel number %s out of range (1 - %d)", word.c_str(), channelNames.size());
+			LogFStatic(L"Channel number %s out of range (1 - %u)", word.c_str(), static_cast<unsigned>(channelNames.size()));
 			channelIndex = -1;
 		}
 	}
